+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/completion/KeywordCompletionProvider.java
@@ -1,15 +1,5 @@
 package dev.xeonkryptos.xeonrobotframeworkplugin.ide.completion;
 
-import dev.xeonkryptos.xeonrobotframeworkplugin.util.LookupElementUtil;
-import dev.xeonkryptos.xeonrobotframeworkplugin.ide.config.RobotOptionsProvider;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.dto.ImportType;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.DefinedKeyword;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.DefinedParameter;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.Heading;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordFile;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotFile;
-import dev.xeonkryptos.xeonrobotframeworkplugin.util.GlobalConstants;
-import com.intellij.codeInsight.TailType;
 import com.intellij.codeInsight.TailTypes;
 import com.intellij.codeInsight.completion.CompletionParameters;
 import com.intellij.codeInsight.completion.CompletionProvider;
@@ -18,66 +8,84 @@ import com.intellij.codeInsight.lookup.LookupElement;
 import com.intellij.codeInsight.lookup.LookupElementBuilder;
 import com.intellij.codeInsight.lookup.TailTypeDecorator;
 import com.intellij.icons.AllIcons.Nodes;
-import com.intellij.openapi.editor.Document;
-import com.intellij.openapi.editor.Editor;
-import com.intellij.openapi.util.TextRange;
+import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
+import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.util.ProcessingContext;
+import dev.xeonkryptos.xeonrobotframeworkplugin.ide.config.RobotOptionsProvider;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.dto.ImportType;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.dto.ParameterDto;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.DefinedKeyword;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.DefinedParameter;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.Heading;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordFile;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordStatement;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotFile;
+import dev.xeonkryptos.xeonrobotframeworkplugin.util.LookupElementUtil;
 import org.apache.commons.text.WordUtils;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
+import java.util.stream.Collectors;
 
 class KeywordCompletionProvider extends CompletionProvider<CompletionParameters> {
 
     @Override
     protected void addCompletions(@NotNull CompletionParameters parameters, @NotNull ProcessingContext context, @NotNull CompletionResultSet result) {
         if (CompletionProviderUtils.isIndexPositionAWhitespaceCharacter(parameters)) {
-            Heading heading = CompletionProviderUtils.getHeading(parameters.getPosition());
+            PsiElement position = parameters.getPosition();
+            Heading heading = CompletionProviderUtils.getHeading(position);
             if (heading != null && (heading.containsTestCases() || heading.containsKeywordDefinitions() || heading.containsTasks())) {
-                KeywordCompletionModification keywordCompletionModification = Arrays.stream(KeywordCompletionModification.values())
-                                                                                    .filter(modification -> modification.getIdentifier() != null)
-                                                                                    .filter(modification -> isStartingWithCharacter(parameters,
-                                                                                                                                    modification.getIdentifier()))
-                                                                                    .findFirst()
-                                                                                    .orElse(KeywordCompletionModification.NONE);
-                addDefinedKeywordsFromFile(result, parameters.getOriginalFile(), keywordCompletionModification);
+                PsiElement positionContext = position.getContext();
+                KeywordStatement keywordStatement = PsiTreeUtil.getParentOfType(positionContext, KeywordStatement.class);
+                assert keywordStatement != null;
+
+                String name = keywordStatement.getName();
+                KeywordCompletionModification keywordCompletionModification = KeywordCompletionModification.NONE;
+                if (!name.isEmpty()) {
+                    char firstCharacter = name.charAt(0);
+                    keywordCompletionModification = Arrays.stream(KeywordCompletionModification.values())
+                                                          .filter(modification -> modification.getIdentifier() != null)
+                                                          .filter(modification -> firstCharacter == modification.getIdentifier())
+                                                          .findFirst()
+                                                          .orElse(KeywordCompletionModification.NONE);
+                }
+                Collection<DefinedParameter> alreadyAddedParameters = keywordStatement.getParameters()
+                                                                                      .stream()
+                                                                                      .map(param -> new ParameterDto(param, param.getParameterName(), null))
+                                                                                      .collect(Collectors.toSet());
+
+                addDefinedKeywordsFromFile(result, parameters.getOriginalFile(), keywordCompletionModification, alreadyAddedParameters);
             }
         }
     }
 
-    private boolean isStartingWithCharacter(@NotNull CompletionParameters parameters, char modificationCharacter) {
-        int offset = parameters.getOffset();
-
-        Document document = parameters.getEditor().getDocument();
-        int lineNumber = document.getLineNumber(offset);
-        int lineStartOffset = document.getLineStartOffset(lineNumber);
-        String textBeforeOffset = document.getText(new TextRange(lineStartOffset, offset));
-        if (textBeforeOffset.trim().isEmpty()) {
-            return false;
-        }
-        int firstCharacterInLine = textBeforeOffset.trim().codePointAt(0);
-        return modificationCharacter == firstCharacterInLine;
-    }
-
-    private void addDefinedKeywordsFromFile(CompletionResultSet resultSet, PsiFile file, KeywordCompletionModification keywordCompletionModification) {
+    private void addDefinedKeywordsFromFile(CompletionResultSet resultSet,
+                                            PsiFile file,
+                                            KeywordCompletionModification keywordCompletionModification,
+                                            Collection<DefinedParameter> alreadyAddedParameters) {
         if (file instanceof RobotFile robotFile) {
-            boolean capitalizeKeywords = RobotOptionsProvider.getInstance(robotFile.getProject()).capitalizeKeywords();
-            addDefinedKeywords(robotFile.getDefinedKeywords(), resultSet, capitalizeKeywords, keywordCompletionModification).forEach(lookupElement -> {
-                lookupElement.putUserData(CompletionKeys.ROBOT_LOOKUP_CONTEXT, RobotLookupContext.KEYWORDS);
-                lookupElement.putUserData(CompletionKeys.ROBOT_LOOKUP_ELEMENT_TYPE, RobotLookupElementType.KEYWORD);
-            });
-            boolean allowTransitiveImports = RobotOptionsProvider.getInstance(file.getProject()).allowTransitiveImports();
+            RobotOptionsProvider robotOptionsProvider = RobotOptionsProvider.getInstance(robotFile.getProject());
+            boolean capitalizeKeywords = robotOptionsProvider.capitalizeKeywords();
+            boolean allowTransitiveImports = robotOptionsProvider.allowTransitiveImports();
+
+            Collection<DefinedKeyword> definedKeywordsFromRobotFile = robotFile.getDefinedKeywords();
+            addDefinedKeywords(definedKeywordsFromRobotFile, resultSet, capitalizeKeywords, keywordCompletionModification, alreadyAddedParameters).forEach(
+                    lookupElement -> {
+                        lookupElement.putUserData(CompletionKeys.ROBOT_LOOKUP_CONTEXT, RobotLookupContext.KEYWORDS);
+                        lookupElement.putUserData(CompletionKeys.ROBOT_LOOKUP_ELEMENT_TYPE, RobotLookupElementType.KEYWORD);
+                    });
             for (KeywordFile importedFile : robotFile.getImportedFiles(allowTransitiveImports)) {
                 if (importedFile.getImportType() != ImportType.VARIABLES) {
-                    addDefinedKeywords(importedFile.getDefinedKeywords(),
+                    Collection<DefinedKeyword> definedKeywordsFromImportedFile = importedFile.getDefinedKeywords();
+                    addDefinedKeywords(definedKeywordsFromImportedFile,
                                        resultSet,
                                        capitalizeKeywords,
-                                       keywordCompletionModification).forEach(lookupElement -> {
+                                       keywordCompletionModification,
+                                       alreadyAddedParameters).forEach(lookupElement -> {
                         lookupElement.putUserData(CompletionKeys.ROBOT_LOOKUP_CONTEXT, RobotLookupContext.KEYWORDS);
                         lookupElement.putUserData(CompletionKeys.ROBOT_LOOKUP_ELEMENT_TYPE, RobotLookupElementType.KEYWORD);
                     });
@@ -89,7 +97,8 @@ class KeywordCompletionProvider extends CompletionProvider<CompletionParameters>
     private Collection<LookupElement> addDefinedKeywords(Collection<DefinedKeyword> keywords,
                                                          CompletionResultSet resultSet,
                                                          boolean capitalize,
-                                                         KeywordCompletionModification keywordCompletionModification) {
+                                                         KeywordCompletionModification keywordCompletionModification,
+                                                         Collection<DefinedParameter> alreadyAddedParameters) {
         List<LookupElement> lookupElements = new ArrayList<>();
         for (DefinedKeyword keyword : keywords) {
             String keywordName = keyword.getKeywordName();
@@ -110,7 +119,7 @@ class KeywordCompletionProvider extends CompletionProvider<CompletionParameters>
 
             TailTypeDecorator<LookupElementBuilder> tailTypeDecorator;
             if (keyword.hasParameters()) {
-                tailTypeDecorator = keywordCompletionModification.createTail(decoratedElement, keyword);
+                tailTypeDecorator = keywordCompletionModification.createTail(decoratedElement, keyword, alreadyAddedParameters);
             } else {
                 tailTypeDecorator = TailTypeDecorator.withTail(decoratedElement, TailTypes.noneType());
             }
@@ -118,96 +127,5 @@ class KeywordCompletionProvider extends CompletionProvider<CompletionParameters>
             lookupElements.add(tailTypeDecorator);
         }
         return lookupElements;
-    }
-
-    private enum KeywordCompletionModification {
-        NONE {
-            @Nullable
-            @Override
-            public Character getIdentifier() {
-                return null;
-            }
-
-            @Override
-            public TailTypeDecorator<LookupElementBuilder> createTail(LookupElementBuilder decoratedElement, DefinedKeyword keyword) {
-                return TailTypeDecorator.withTail(decoratedElement, TailTypes.noneType());
-            }
-        }, ONLY_MANDATORY {
-            @NotNull
-            @Override
-            public Character getIdentifier() {
-                return '*';
-            }
-
-            @Override
-            public TailTypeDecorator<LookupElementBuilder> createTail(LookupElementBuilder decoratedElement, DefinedKeyword keyword) {
-                return TailTypeDecorator.withTail(decoratedElement, new TailType() {
-                    @Override
-                    public int processTail(Editor editor, int tailOffset) {
-                        String keywordName = keyword.getKeywordName();
-                        Document document = editor.getDocument();
-
-                        int lineNumber = document.getLineNumber(tailOffset);
-                        int lineStartOffset = document.getLineStartOffset(lineNumber);
-                        int keywordStartOffset = tailOffset - keywordName.length();
-                        String spaceBeforeOffset = document.getText(new TextRange(lineStartOffset, keywordStartOffset));
-                        String prefixIndentationText = "\n" + spaceBeforeOffset + GlobalConstants.ELLIPSIS + GlobalConstants.DEFAULT_INDENTATION;
-
-                        int currentOffset = tailOffset;
-                        int addedOffset = 0;
-                        for (DefinedParameter parameter : keyword.getParameters()) {
-                            if (!parameter.hasDefaultValue()) {
-                                String parameterInsertString = prefixIndentationText + parameter.getLookup() + "=";
-                                document.insertString(currentOffset, parameterInsertString);
-                                currentOffset += parameterInsertString.length();
-                                addedOffset += parameterInsertString.length();
-                            }
-                        }
-                        return moveCaret(editor, tailOffset, addedOffset);
-                    }
-                });
-            }
-        }, ALL {
-            @NotNull
-            @Override
-            public Character getIdentifier() {
-                return '/';
-            }
-
-            @Override
-            public TailTypeDecorator<LookupElementBuilder> createTail(LookupElementBuilder decoratedElement, DefinedKeyword keyword) {
-                return TailTypeDecorator.withTail(decoratedElement, new TailType() {
-                    @Override
-                    public int processTail(Editor editor, int tailOffset) {
-                        String keywordName = keyword.getKeywordName();
-                        Document document = editor.getDocument();
-
-                        int lineNumber = document.getLineNumber(tailOffset);
-                        int lineStartOffset = document.getLineStartOffset(lineNumber);
-                        int keywordStartOffset = tailOffset - keywordName.length();
-                        String spaceBeforeOffset = document.getText(new TextRange(lineStartOffset, keywordStartOffset));
-                        String prefixIndentationText = "\n" + spaceBeforeOffset + GlobalConstants.ELLIPSIS + GlobalConstants.DEFAULT_INDENTATION;
-
-                        int currentOffset = tailOffset;
-                        int addedOffset = 0;
-                        for (DefinedParameter parameter : keyword.getParameters()) {
-                            String parameterInsertString = prefixIndentationText + parameter.getLookup() + "=";
-                            if (parameter.hasDefaultValue()) {
-                                parameterInsertString += parameter.getDefaultValue();
-                            }
-                            document.insertString(currentOffset, parameterInsertString);
-                            currentOffset += parameterInsertString.length();
-                            addedOffset += parameterInsertString.length();
-                        }
-                        return moveCaret(editor, tailOffset, addedOffset);
-                    }
-                });
-            }
-        };
-
-        @Nullable
-        public abstract Character getIdentifier();
-
-        public abstract TailTypeDecorator<LookupElementBuilder> createTail(LookupElementBuilder decoratedElement, DefinedKeyword keyword);
     }
 }